refactor(core): extract inline collection into InlineCollector
Move the inline schema/enum collection (walk, dedup, name, model) out of
ModelGenerator into InlineCollector, alongside the resolution half. The
generic pipeline is now collectInlineSchemas/collectInlineEnums building on
inlineRefs + toNamedModels. ModelGenerator stays focused on KotlinPoet
generation (729 -> 634 lines).
